Refactors types conversion
