+++ bindings/R/FortMyrmidon/src/queries.cpp
@@ -1,5 +1,7 @@
 #include <FortMyrmidon/queries.h>
 
+#include "types.h"
+
 #include "time.h"
 #include "experiment.h"
 #include "matchers.h"
@@ -12,270 +14,8 @@
 
 using namespace fort::myrmidon;
 
-#define Vectorize(...) {__VA_ARGS__}
-#define CreateDataFrameOptimized(VarName,elements,names,nrows ) \
-	Rcpp::List VarName(elements); \
-	VarName.attr("names") = Rcpp::CharacterVector(names); \
-	VarName.attr("row.names") = Rcpp::IntegerVector::create(NA_INTEGER,(nrows)); \
-	VarName.attr("class") = "data.frame"
 
 
-std::tuple<Rcpp::IntegerVector,Rcpp::NumericVector,Rcpp::NumericVector,Rcpp::NumericVector>
-fmPositionData_asR(const PositionedAntList & ants) {
-		size_t n = ants.size();
-		Rcpp::IntegerVector IDs(n);
-		Rcpp::NumericVector X(n),Y(n),Angle(n);
-		size_t i = 0;
-		for ( const auto & ant : ants ) {
-			IDs[i] = ant.ID;
-			X[i] = ant.Position.x();
-			Y[i] = ant.Position.y();
-			Angle[i] = ant.Angle;
-			++i;
-		}
-		return {IDs,X,Y,Angle};
-}
-
-SEXP fmIdentifiedFrames_asR(std::vector<IdentifiedFrame::ConstPtr> & frames, bool withZone ) {
-	using namespace Rcpp;
-	List positions(frames.size());
-	DatetimeVector time(frames.size());
-	IntegerVector width(frames.size()),height(frames.size()),space(frames.size());
-	size_t i = 0;
-	for ( auto & f : frames ) {
-		auto [IDs,X,Y,Angle] = fmPositionData_asR(f->Positions);
-		if ( withZone == true ) {
-			CreateDataFrameOptimized(df,
-			                         Vectorize(IDs,X,Y,Angle, IntegerVector(f->Zones.begin(),f->Zones.end())),
-			                         Vectorize("antID","x","y","angle","zone"),
-			                         IDs.size());
-
-			positions[i] = df;
-		} else {
-			CreateDataFrameOptimized(df,
-			                         Vectorize(IDs,X,Y,Angle),
-			                         Vectorize("antID","x","y","angle"),
-			                         IDs.size());
-			positions[i] = df;
-		}
-
-		time[i] = fmTime_asR(f->FrameTime);
-		width[i] = f->Width;
-		height[i] = f->Height;
-		space[i] = f->Space;
-		++i;
-		f.reset();
-	}
-
-	return List::create(_["frames"] = DataFrame::create(_["time"] = time,
-	                                                    _["space"] = space,
-	                                                    _["width"] = width,
-	                                                    _["height"] = height),
-	                    _["positions"] = positions);
-}
-
-
-std::string fmInteractionTypes_asStr(const InteractionTypes & types ) {
-	std::ostringstream res;
-	for ( size_t i = 0; i < types.rows(); ++i ) {
-		res << types(i,0) <<  "-" <<  types(i,1) << ",";
-	}
-	auto ress = res.str();
-	ress.pop_back();
-	return ress;
-}
-
-SEXP fmCollisionData_asR(std::vector<Query::CollisionData> & frames ) {
-	using namespace Rcpp;
-	List positions(frames.size());
-	DatetimeVector time(frames.size());
-	IntegerVector width(frames.size()),height(frames.size()),space(frames.size());
-
-	std::vector<std::string> types;
-	std::vector<AntID> ant1IDs,ant2IDs;
-	std::vector<size_t> frameRowIndex;
-	std::vector<ZoneID> zone;
-	size_t i = 0;
-	for ( auto & [identified,collided] : frames) {
-		auto [IDs,X,Y,Angle] = fmPositionData_asR(identified->Positions);
-		CreateDataFrameOptimized(df,
-		                         Vectorize(IDs,X,Y,Angle, IntegerVector(identified->Zones.begin(),identified->Zones.end())),
-		                         Vectorize("antID","x","y","angle","zone"),
-		                         IDs.size());
-
-		positions[i] = df;
-
-		time[i] = fmTime_asR(identified->FrameTime);
-		width[i] = identified->Width;
-		height[i] = identified->Height;
-		space[i] = identified->Space;
-
-		for ( const auto & collision : collided->Collisions ) {
-			frameRowIndex.push_back(i+1); // R index starts at 1, not 0.
-			ant1IDs.push_back(collision.IDs.first);
-			ant2IDs.push_back(collision.IDs.second);
-			zone.push_back(collision.Zone);
-			types.push_back(fmInteractionTypes_asStr(collision.Types));
-		}
-		++i;
-		identified.reset();
-		collided.reset();
-	}
-	return List::create(_["frames"] = DataFrame::create(_["time"] = time,
-	                                                    _["space"] = space,
-	                                                    _["width"] = width,
-	                                                    _["height"]  = height),
-	                    _["positions"] = positions,
-	                    _["collisions"] = DataFrame::create(_["ant1"] = IntegerVector(ant1IDs.begin(),ant1IDs.end()),
-	                                                        _["ant2"] = IntegerVector(ant2IDs.begin(),ant2IDs.end()),
-	                                                        _["zone"] = IntegerVector(zone.begin(),zone.end()),
-	                                                        _["types"] = CharacterVector(types.cbegin(),types.cend()),
-	                                                        _["frames_row_index"] = IntegerVector(frameRowIndex.begin(),
-		                     frameRowIndex.end())));
-}
-
-Rcpp::DataFrame fmAntTrajectoryPosition_asR(const AntTrajectory::ConstPtr & trajectory) {
-	using namespace Rcpp;
-	size_t nPoints = trajectory->Positions.rows();
-#define numericVectorFromEigen(Var,matrix,col,size) NumericVector Var(&((matrix)(0,col)),&((matrix)(0,col))+size)
- 	numericVectorFromEigen(time,trajectory->Positions,0,nPoints);
-	numericVectorFromEigen(x,trajectory->Positions,1,nPoints);
- 	numericVectorFromEigen(y,trajectory->Positions,2,nPoints);
- 	numericVectorFromEigen(angle,trajectory->Positions,3,nPoints);
-	if ( trajectory->Zones.empty() == false ) {
-		CreateDataFrameOptimized(df,
-		                         Vectorize(time,x,y,angle, IntegerVector(trajectory->Zones.cbegin(),trajectory->Zones.cend())),
-		                         Vectorize("time","x","y","angle","zone"),
-		                         trajectory->Positions.rows());
-		return df;
-	}
-	CreateDataFrameOptimized(df,
-	                         Vectorize(time,x,y,angle),
-	                         Vectorize("time","x","y","angle"),
-	                         trajectory->Positions.rows());
-
-	return df;
-}
-
-
-std::tuple<Rcpp::DataFrame,Rcpp::List>
-fmAntTrajectories_asR(const std::vector<AntTrajectory::ConstPtr> & trajectories) {
-	using namespace Rcpp;
-	auto n = trajectories.size();
-	IntegerVector antID(n),space(n);
-	DatetimeVector start(n);
-	List positions(n);
-	size_t i = 0;
-	for ( const auto & trajectory : trajectories ) {
-		antID[i] = trajectory->Ant;
-		space[i] = trajectory->Space;
-		start[i] = fmTime_asR(trajectory->Start);
-		positions[i] = fmAntTrajectoryPosition_asR(trajectory);
-		++i;
-	}
-
-	return {DataFrame::create(_["antID"] = antID,
-	                          _["start"] = start,
-	                          _["space"] = space),
-	        positions};
-
-}
-
-std::tuple<Rcpp::DataFrame,Rcpp::List,Rcpp::List>
-fmAntInteractions_asR(std::vector<AntInteraction::ConstPtr> & interactions) {
-	using namespace Rcpp;
-	auto n = interactions.size();
-	List ant1Trajectory(n),ant2Trajectory(n);
-	IntegerVector ant1(n),ant2(n),space(n);
-	DatetimeVector start(n),end(n);
-	CharacterVector types(n);
-
-	size_t i = 0;
-	for ( auto & interaction : interactions ) {
-		ant1[i] = interaction->IDs.first;
-		ant2[i] = interaction->IDs.second;
-		start[i] = fmTime_asR(interaction->Start);
-		end[i] = fmTime_asR(interaction->End);
-		space[i] = interaction->Space;
-
-		types[i] = fmInteractionTypes_asStr(interaction->Types);
-
-		ant1Trajectory[i] = fmAntTrajectoryPosition_asR(interaction->Trajectories.first);
-		ant2Trajectory[i] = fmAntTrajectoryPosition_asR(interaction->Trajectories.second);
-
-		++i;
-		interaction.reset();
-	}
-
-	return {DataFrame::create(_["ant1"] = ant1,
-	                          _["ant2"] = ant2,
-	                          _["start"] = start,
-	                          _["end"] = end,
-	                          _["space"] = space,
-	                          _["types"]  = types),
-	        ant1Trajectory,
-	        ant2Trajectory};
-}
-
-
-Rcpp::DataFrame
-fmAntInteractionsSummarized_asR(std::vector<AntInteraction::ConstPtr> & interactions,
-                               std::vector<double> & meanAnt1,
-                               std::vector<double> & meanAnt2,
-                                const std::vector<std::string> & zoneAnt1,
-                                const std::vector<std::string> & zoneAnt2) {
-	using namespace Rcpp;
-	auto n = interactions.size();
-
-	IntegerVector ant1(n),ant2(n),space(n);
-	DatetimeVector start(n),end(n);
-	CharacterVector types(n);
-
-	size_t i = 0;
-	for ( auto & interaction : interactions) {
-		ant1[i] = interaction->IDs.first;
-		ant2[i] = interaction->IDs.second;
-		start[i] = fmTime_asR(interaction->Start);
-		end[i] = fmTime_asR(interaction->End);
-		space[i] = interaction->Space;
-
-		types[i] = fmInteractionTypes_asStr(interaction->Types);
-
-		++i;
-		interaction.reset();
-	}
-
-	//first go from row-major to column-major using eigen.
-	Eigen::MatrixXd mean1 = Eigen::Map<Eigen::Matrix<double,Eigen::Dynamic,3,Eigen::RowMajor>>(&meanAnt1[0],n,3);
-	meanAnt1.clear();
-	Eigen::MatrixXd mean2 = Eigen::Map<Eigen::Matrix<double,Eigen::Dynamic,3,Eigen::RowMajor>>(&meanAnt2[0],n,3);
-	meanAnt2.clear();
-
-	// now duplicates again the data to R, from column major order
-	numericVectorFromEigen(x1mean,mean1,0,n);
-	numericVectorFromEigen(y1mean,mean1,1,n);
-	numericVectorFromEigen(angle1mean,mean1,2,n);
-
-	numericVectorFromEigen(x2mean,mean2,0,n);
-	numericVectorFromEigen(y2mean,mean2,1,n);
-	numericVectorFromEigen(angle2mean,mean2,2,n);
-
-	// now we can put everything in data.frame
-	return DataFrame::create(_["ant1"] = ant1,
-	                         _["ant2"] = ant2,
-	                         _["start"] = start,
-	                         _["end"] = end,
-	                         _["space"] = space,
-	                         _["types"] = types,
-	                         _["ant1_x_mean"] = x1mean,
-	                         _["ant1_y_mean"] = y1mean,
-	                         _["ant1_angle_mean"] = angle1mean,
-	                         _["ant2_x_mean"] = x2mean,
-	                         _["ant2_y_mean"] = y2mean,
-	                         _["ant2_angle_mean"] = angle2mean,
-	                         _["ant1_zones"] = CharacterVector(zoneAnt1.cbegin(),zoneAnt1.cend()),
-	                         _["ant2_zones"] = CharacterVector(zoneAnt2.cbegin(),zoneAnt2.cend()));
-}
 
 Rcpp::DataFrame fmQuery_computeMeasurementFor(const CExperiment & experiment,
                                               AntID antID,
@@ -440,8 +180,13 @@ SEXP fmQueryIdentifyFrames(const fort::myrmidon::CExperiment & experiment,
 	                             endTime,
 	                             computeZones,
 	                             true);
-	SET_PD("R conversion");
-	auto res = fmIdentifiedFrames_asR(res_,computeZones);
+	SET_PD("R conversion");\
+	SEXP res;
+	if ( computeZones == true) {
+		res = fmIdentifiedFramesWithZones_asR(res_);
+	} else {
+		res = fmIdentifiedFrames_asR(res_);
+	}
 	CLEAR_PD();
 	return res;
 }
@@ -505,8 +250,8 @@ double meanUS(const std::vector<Duration> & durations) {
 }
 
 
-std::pair<Eigen::Vector3d,std::string> SummarizeTrajectory(const AntTrajectory::ConstPtr & trajectory) {
-	std::set<ZoneID> set(trajectory->Zones.begin(),trajectory->Zones.end());
+std::pair<Eigen::Vector3d,std::string> SummarizeTrajectory(const AntTrajectorySegment & segment) {
+	std::set<ZoneID> set(segment.Trajectory->Zones.begin(),segment.Trajectory->Zones.end());
 	std::ostringstream oss;
 	for ( const auto & zone : set ) {
 		oss << zone << ",";
@@ -514,7 +259,7 @@ std::pair<Eigen::Vector3d,std::string> SummarizeTrajectory(const AntTrajectory::
 	std::string zones(oss.str());
 	zones.pop_back();
 
-	return {trajectory->Positions.block(0,1,trajectory->Positions.rows(),3).colwise().mean(),
+	return {segment.Trajectory->Positions.block(segment.Begin,1,segment.End-segment.Begin,3).colwise().mean(),
 	        zones};
 }
 
@@ -529,8 +274,8 @@ SEXP fmQueryComputeAntInteractions(const CExperiment & experiment,
 
 	std::vector<AntTrajectory::ConstPtr> resTrajectories;
 	std::vector<AntInteraction::ConstPtr>  resInteractions;
-	std::vector<double> meanAnt1,meanAnt2;
-	std::vector<std::string> zone1,zone2;
+	TrajectorySummary ant1Summary,ant2Summary;
+	TrajectoryIndexing ant1Indexing,ant2Indexing;
 
 	DECLARE_PD();
 
@@ -555,7 +300,7 @@ SEXP fmQueryComputeAntInteractions(const CExperiment & experiment,
 			resInteractions.push_back(interaction);
 		};
 
-	if ( reportLocalTrajectories == false ) {
+	if ( reportTrajectories == false ) {
 		storeInteractions =
 			[&](const AntInteraction::ConstPtr & interaction) {
 				auto smallerInteraction = std::make_shared<AntInteraction>(AntInteraction{
@@ -571,15 +316,8 @@ SEXP fmQueryComputeAntInteractions(const CExperiment & experiment,
 				auto [m1,z1] = SummarizeTrajectory(interaction->Trajectories.first);
 				auto [m2,z2] = SummarizeTrajectory(interaction->Trajectories.second);
 
-				meanAnt1.push_back(m1.x());
-				meanAnt1.push_back(m1.y());
-				meanAnt1.push_back(m1.z());
-				meanAnt2.push_back(m2.x());
-				meanAnt2.push_back(m2.y());
-				meanAnt2.push_back(m2.z());
-
-				zone1.push_back(z1);
-				zone2.push_back(z2);
+				ant1Summary.Push(m1,z1);
+				ant2Summary.Push(m2,z2);
 
 			};
 	}
@@ -593,7 +331,7 @@ SEXP fmQueryComputeAntInteractions(const CExperiment & experiment,
 	                                     matcher,
 	                                     singleThreaded);
 	Rcpp::List res;
-	if (reportGlobalTrajectories == true ) {
+	if (reportTrajectories == true ) {
 		SET_PD("R trajectory conversion");
 		auto [summary,trajectories] = fmAntTrajectories_asR(resTrajectories);
 		res["summaryTrajectory"] = summary;
@@ -601,17 +339,14 @@ SEXP fmQueryComputeAntInteractions(const CExperiment & experiment,
 	}
 
 	SET_PD("R interaction conversion");
-	if (reportLocalTrajectories == true ) {
-		auto [interactions,t1,t2] = fmAntInteractions_asR(resInteractions);
-		res["interactions"] = interactions;
-		res["ant1Trajectory"] = t1;
-		res["ant2Trajectory"] = t2;
+	if (reportTrajectories == true ) {
+		res["interactions"] = fmAntInteractions_asR(resInteractions,
+		                                            ant1Indexing,
+		                                            ant2Indexing);
 	} else {
 		res["interactions"] = fmAntInteractionsSummarized_asR(resInteractions,
-		                                                      meanAnt1,
-		                                                      meanAnt2,
-		                                                      zone1,
-		                                                      zone2);
+		                                                      ant1Summary,
+		                                                      ant2Summary);
 	}
 	CLEAR_PD();
 	return res;
