+++ bindings/R/test.R
@@ -1,7 +1,7 @@
 devtools::load_all("FortMyrmidon");
 e <- fmExperimentOpenReadOnly("/home/atuleu/Documents/large.myrmidon");
 start <- fmTimeParse('2020-05-14T08:00:00.000Z');
-end <- fmTimeParse('2020-05-14T18:01:00.000Z');
+end <- fmTimeParse('2020-05-14T08:01:00.000Z');
 gap <- fmSecond(3);
 dd <- fmQueryComputeAntInteractions(e,
                                     start = start,
@@ -9,6 +9,5 @@ dd <- fmQueryComputeAntInteractions(e,
                                     maximumGap = fmSecond(20),
                                     singleThreaded = FALSE,
                                     showProgress = TRUE,
-                                    reportGlobalTrajectories = TRUE,
-                                    reportLocalTrajectories = TRUE);
+                                    reportTrajectories = TRUE);
 #dd$interactions
